@@ -13,7 +13,7 @@
         }
     ],
     "require": {
-        "php": "~5.3",
+        "php": "~5.4",
         "silex/silex": "~1.1",
         "twig/twig": "~1.12",
         "symfony/twig-bridge": "~2.4",
